Securevault/requirements.txt
Django>=5.0,<6.0
python-decouple>=3.8
cryptography>=42.0
pyotp>=2.9.0
pyclamd>=0.4.0

